Safari-609.1.11.1/LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-up-click-over-media-does-not-dimiss-media-controls-when-media-is-playing-expected.txt

Clicking over the media when the tracks panel is up and the media is playing should dismiss the tracks panel but not the media controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media started playing.

PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
=> Tracks button is visible.

Pressing on the tracks button.
PASS shadowRoot.querySelector('.tracks-panel') became different from null
=> Tracks panel is visible.

Clicking over the media to dismiss the tracks panel.
PASS shadowRoot.querySelector('.media-controls').classList.contains('faded') is false
PASS shadowRoot.querySelector('.tracks-panel').classList.contains('fade-out') is true

PASS successfullyParsed is true

TEST COMPLETE
